# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Whatsgo - whatsmeow",
        "title": "Whatsgo",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/send/contact": {
            "post": {
                "description": "Send a contact message",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Send Message"
                ],
                "summary": "Send a contact message",
                "parameters": [
                    {
                        "description": "Message data",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/whatsgo_pkg_sendMessage_service.ContactStruct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Error on validation",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/send/link": {
            "post": {
                "description": "Send a link message",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Send Message"
                ],
                "summary": "Send a link message",
                "parameters": [
                    {
                        "description": "Message data",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/whatsgo_pkg_sendMessage_service.LinkStruct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Error on validation",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/send/location": {
            "post": {
                "description": "Send a location message",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Send Message"
                ],
                "summary": "Send a location message",
                "parameters": [
                    {
                        "description": "Message data",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/whatsgo_pkg_sendMessage_service.LocationStruct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Error on validation",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/send/media": {
            "post": {
                "description": "Send a media message",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Send Message"
                ],
                "summary": "Send a media message",
                "parameters": [
                    {
                        "description": "Message data",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/whatsgo_pkg_sendMessage_service.MediaStruct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Error on validation",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/send/poll": {
            "post": {
                "description": "Send a poll message",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Send Message"
                ],
                "summary": "Send a poll message",
                "parameters": [
                    {
                        "description": "Message data",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/whatsgo_pkg_sendMessage_service.PollStruct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Error on validation",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/send/sticker": {
            "post": {
                "description": "Send a sticker message",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Send Message"
                ],
                "summary": "Send a sticker message",
                "parameters": [
                    {
                        "description": "Message data",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/whatsgo_pkg_sendMessage_service.StickerStruct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Error on validation",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/send/text": {
            "post": {
                "description": "Send a text message",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Send Message"
                ],
                "summary": "Send a text message",
                "parameters": [
                    {
                        "description": "Message data",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/whatsgo_pkg_sendMessage_service.TextStruct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Error on validation",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "gin.H": {
            "type": "object",
            "additionalProperties": {}
        },
...
        "whatsgo_pkg_sendMessage_service.ContactStruct": {
            "type": "object",
            "properties": {
                "delay": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "mentionAll": {
                    "type": "boolean"
                },
                "mentionedJid": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "quoted": {
                    "$ref": "#/definitions/whatsgo_pkg_sendMessage_service.QuotedStruct"
                },
                "vcard": {
                    "$ref": "#/definitions/whatsgo_pkg_utils.VCardStruct"
                }
            }
        },
        "whatsgo_pkg_sendMessage_service.LinkStruct": {
            "type": "object",
            "properties": {
                "delay": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "imgUrl": {
                    "type": "string"
                },
                "mentionAll": {
                    "type": "boolean"
                },
                "mentionedJid": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "quoted": {
                    "$ref": "#/definitions/whatsgo_pkg_sendMessage_service.QuotedStruct"
                },
                "text": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "whatsgo_pkg_sendMessage_service.LocationStruct": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "delay": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "mentionAll": {
                    "type": "boolean"
                },
                "mentionedJid": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "quoted": {
                    "$ref": "#/definitions/whatsgo_pkg_sendMessage_service.QuotedStruct"
                }
            }
        },
        "whatsgo_pkg_sendMessage_service.MediaStruct": {
            "type": "object",
            "properties": {
                "caption": {
                    "type": "string"
                },
                "delay": {
                    "type": "integer"
                },
                "filename": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "mentionAll": {
                    "type": "boolean"
                },
                "mentionedJid": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "quoted": {
                    "$ref": "#/definitions/whatsgo_pkg_sendMessage_service.QuotedStruct"
                },
                "type": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "whatsgo_pkg_sendMessage_service.PollStruct": {
            "type": "object",
            "properties": {
                "delay": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "maxAnswer": {
                    "type": "integer"
                },
                "mentionAll": {
                    "type": "boolean"
                },
                "mentionedJid": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "question": {
                    "type": "string"
                },
                "quoted": {
                    "$ref": "#/definitions/whatsgo_pkg_sendMessage_service.QuotedStruct"
                }
            }
        },
        "whatsgo_pkg_sendMessage_service.QuotedStruct": {
            "type": "object",
            "properties": {
                "messageId": {
                    "type": "string"
                },
                "participant": {
                    "type": "string"
                }
            }
        },
        "whatsgo_pkg_sendMessage_service.StickerStruct": {
            "type": "object",
            "properties": {
                "delay": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "mentionAll": {
                    "type": "boolean"
                },
                "mentionedJid": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "quoted": {
                    "$ref": "#/definitions/whatsgo_pkg_sendMessage_service.QuotedStruct"
                },
                "sticker": {
                    "type": "string"
                }
            }
        },
        "whatsgo_pkg_sendMessage_service.TextStruct": {
            "type": "object",
            "properties": {
                "delay": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "mentionAll": {
                    "type": "boolean"
                },
                "mentionedJid": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "quoted": {
                    "$ref": "#/definitions/whatsgo_pkg_sendMessage_service.QuotedStruct"
                },
                "text": {
                    "type": "string"
                }
            }
        },
...
        "whatsgo_pkg_utils.VCardStruct": {
            "type": "object",
            "properties": {
                "fullName": {
                    "type": "string"
                },
                "organization": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
...
    }
}